requirements.txt
streamlit==1.37.1
pdfplumber==0.11.4
pypdf==4.3.1
python-dateutil==2.9.0.post0
psycopg2-binary==2.9.9
reportlab==4.2.2
pydantic==2.8.2
python-dotenv==1.0.1
